Clean up the server API
Split out the Models from the mechanics of the requests.
