+++ packages/_server/package.json
@@ -14,25 +14,15 @@
   },
   "type": "module",
   "main": "dist/main.cjs",
-  "typings": "dist/api.d.cts",
   "exports": {
-    ".": "./dist/main.cjs",
-    "./api": "./dist/api.cjs"
-  },
-  "typesVersions": {
-    "*": {
-      "*.d.cts": [
-        "dist/*.d.cts"
-      ],
-      "api": [
-        "dist/api.d.cts"
-      ],
-      "dist/api.d.cts": [
-        "dist/api.d.cts"
-      ],
-      "*": [
-        "dist/*.d.cts"
-      ]
+    ".": {
+      "import": "./dist/main.cjs",
+      "require": "./dist/main.cjs"
+    },
+    "./api": {
+      "types": "./dist/api.d.cts",
+      "import": "./dist/api.cjs",
+      "require": "./dist/api.cjs"
     }
   },
   "devDependencies": {
@@ -52,14 +42,14 @@
     "cspell-glob": "^7.3.7",
     "cspell-lib": "^7.3.7",
     "gensequence": "^6.0.0",
+    "json-rpc-api": "file:../json-rpc-api",
     "node-watch": "^0.7.4",
     "rxjs": "^7.8.1",
     "utils-disposables": "file:../utils-disposables",
     "vscode-languageserver": "^9.0.1",
     "vscode-languageserver-textdocument": "^1.0.11",
     "vscode-languageserver-types": "^3.17.5",
-    "vscode-uri": "^3.0.8",
-    "vscode-webview-rpc": "file:../webview-rpc"
+    "vscode-uri": "^3.0.8"
   },
   "scripts": {
     "build": "npm run build:esbuild && npm run build:api && npm run build:tsc && npm run build-schema",
