+++ packages/_server/src/api/index.ts
@@ -1,1 +1,3 @@
+export * from './api.js';
 export * from './api.old.js';
+export * from './apiModels.js';
